--- src/Features/Inventory/InventoryPage/index.ts
@@ -1,1 +1,0 @@
-export { default as InventoryPage } from './InventoryPage';
