--- src/Features/Inventory/InventoryTable/index.ts
@@ -1,1 +1,0 @@
-export { default as InventoryTable } from './InventoryTable';
