--- src/components/Cart/index.ts
@@ -1,1 +1,0 @@
-export { default as Cart } from './Cart';
