--- src/components/NavBar/index.ts
@@ -1,1 +1,0 @@
-export { default as NavBar } from './NavBar';
