--- src/components/ProductSearchCriteria/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductSearchCriteria } from './ProductSearchCriteria';
