--- src/pages/Catalog/ProductPage/ProductDetails/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductDetails } from './ProductDetails';
