--- src/pages/Catalog/ProductPage/ProductGallery/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductGallery } from './ProductGallery';
