--- src/pages/Catalog/ProductPage/ProductPurchase/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductPurshase } from './ProductPurchase';
