--- src/pages/Catalog/ProductPage/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductPage } from './ProductPage';
